Rename Index.html to index.html
Change the name of an HTML file, from capital "I" to lower case "i" to fix the link in GitHub. From our GitHub Repositories lesson below the link should work. However it didn’t work. So i decided to change the letter I in html file name. 
"When the page refreshes, scroll back down. In the light blue header section of the “GitHub Pages” section, you will be given a URL. That is the link to your deployed project. You may click on it and see a 404 for the first few minutes or hours depending on how busy GitHub is. Don’t panic. This is why we require you to set up GitHub Pages immediately!"